+++ tests/Transport/NullTransportTest.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Sentry\Tests\Transport;
 
+use GuzzleHttp\Promise\PromiseInterface;
 use PHPUnit\Framework\TestCase;
 use Sentry\Event;
+use Sentry\ResponseStatus;
 use Sentry\Transport\NullTransport;
 
 final class NullTransportTest extends TestCase
@@ -15,6 +17,11 @@ final class NullTransportTest extends TestCase
         $transport = new NullTransport();
         $event = new Event();
 
-        $this->assertSame($event->getId(), $transport->send($event));
+        $promise = $transport->send($event);
+        $promiseResult = $promise->wait();
+
+        $this->assertSame(PromiseInterface::FULFILLED, $promise->getState());
+        $this->assertSame(ResponseStatus::skipped(), $promiseResult->getStatus());
+        $this->assertSame($event, $promiseResult->getEvent());
     }
 }